feat: automate some of the review checks (#4)
This PR adds a script that is able to automatically check some of the
public listing review criteria.

Each criteria is implemented as a separate function, which returns the
checklist item either ticked (it passed) or unticked (it failed or was
unable to be checked).

The intention is that the initial launch of the new public listing
review process automates everything that is reasonable to do with a
moderate amount of effort, and that further improvements will be made in
the future to increase the amount of automation.

In a later PR, this script will be exposed as an entry point so that it
can be used to check how close a charm is to passing review before
actually opening a review request. The specifics of how that entry point
will work are left for that PR.

I intend to add basic unit test coverage for most or all of these
functions *in this PR* - however, I would appreciate an initial review
first to ensure that there aren't going to be significant changes before
implementing the tests.

If there are criteria (for example from [this example
list](https://github.com/tonyandrewmeyer/charmhub-listing-review/issues/10#issuecomment-3181420543))
that you feel should also be automated in the initial launch, please
point those out. Conversely, if you believe that some of this automation
should be dropped, please point that out too.