feat: add issue automation (#5)
This PR adds the core automation of public listing review requests.

The update_issue.py script is run when an issue with the appropriate
label is created or labelled, or when comments are added or edited. The
script:

* Sets a standard summary for the issue.
* Creates instructions and a checklist for the review, and either posts
it as a comment (if one doesn't already exist) or updates the existing
comment with the latest state.
* The above comment also includes assigning the issue (to a manager) and
asking them to delegate it to someone in their team.
* Figures out who the manager has assigned to the review, by looking for
a comment where they have done this (this is more fragile than I would
like, but is also easily fixed by editing a comment, and I don't have
any better ideas).
* Looks for comments by the reviewer and uses those to also update the
state in the updated comment. The reviewer can overrule the automated
checks where necessary.

The raw data for the review (the charm repository location and so forth)
is read from the issue description.

A follow-up PR will make it possible to use this script via an entry
point for checking the status of a charm prior to requesting public
listing, so that authors can get an idea of how close they are. The
output will need to be in a different format so that it's clear what
gets checked automatically and what they need to make their own
judgement call on, and probably we can point to more documentation. The
details of this system are left for that future PR.

Another follow-up PR will improve the way that we track the "done"/"not
done/unknown" status for each item, so that we aren't relying on
string-matching "* [ ]" and "* [x]" (note that currently even using `-`
rather than `*` will break the match). Ideally, that would be in this PR
but it's already as large as I'd like it to be, and it does actually
work as-is, and I'd like to have it actually tested out end-to-end.