fix: temporarily hard-code the CI location for the reviewers file (#12)
Copying this temporary fix from my test repo.

This is definitely not the right fix long term, but I'm not sure where
the reviewers file should live, or the best way for the script to find
it (and maybe it should just use CODEOWNERs or the GitHub list of users
with the right access?).

For now, this lets the workflow run.