feat: only run automation once and simplify getting the reviewer checklist (#17)
Two changes:

* The reviewer's checklist becomes the source-of-truth of the state of
the review. This means we don't re-run the update when comments are
posted or edited (we still do on label change and re-open, to help sort
out errors). We can also remove the complicated "update the bot's list
based on the reviewer's comments" logic, and no longer require the
manager to ping the reviewer in the comments (although that still is
ideal, to get them subscribed and notified).
* Adjustments to the bot's comment to make the review process simpler.
* Put the checklist in backticks so that it's easily copied *as
Markdown*.
* Put the checklist in a "details" block so that the comment doesn't
take as much space by default.
* Move the instruction to the manager to the top of the comment, as
that's the only thing they need to see.
* Remove the "additional checks" section, which was taken from the
second level of the Charm Maturity page, but really doesn't belong here
and just confuses things.
* Remove the hard-coded list of best practices to ignore - we should
just fix the best practices in the docs instead. (This means there are a
couple of duplicates now and a couple that don't make sense, but those
will be fixed in Ops and Charmcraft PRs).

There's an [example issue in my
repo](https://github.com/tonyandrewmeyer/charmhub-listing-review/issues/12)
that has the new comment.

Fixes #16

---------

Co-authored-by: Ben Hoyt <ben.hoyt@canonical.com>