feat: retrieve the best practices list from Ops (#20)
Getting the best practices list is reasonably expensive and adds
complexity, particularly when used outside of CI. We have to clone both
canonical/operator and canonical/charmcraft (and potentially more
repositories in the future), and then parse the doc files to get the
best practices.

We're already doing this work for the unified list of best practices in
the Ops docs. Rather than duplicating that here, just retrieve the list
from the Ops repository.

In #9, the suggestion is to cache the processed best practices list in
this repo, rather than getting it every time. That was suggested before
Ops was also caching it. I don't think we need to have it copied to this
repo as well - one HTTP request is pretty lightweight and avoids the
overhead of having CI keep the copy in sync.

Drive-by fixes:
* If there are no options or no actions, then the automated check that
the names are formatted correctly should pass, not fail.
* `tox -e format`, missed in some previous work, and missing (#8) from
CI.
* Updating the wording of some of the best practices in a recent PR to
Ops, to fix the automation.

Fixes #9

---------

Co-authored-by: James Garner <james.garner@canonical.com>