feat: add the initial version of the self-review tool (#24)
This PR adds the self-review tool, which charmers can use before they
open a review request to see how they are doing. The automated checks
are run, and the full checklist output, so they can see what they'll
need to manually evaluate as well.

Also, since we're pushing to PyPI with this change, add some more of the
basics:
* Our standard code of conduct.
* Our standard security policy
* A changelog
* Enable dependabot
* Enable Zizmor

The best practice source is also fixed - it was correct when I tested in
the `tonyandrewmeyer` repository, but apparently just changing the org
to `canonical` isn't correct, it's quite a different location.

Some of the remaining ugliness will be fixed by existing issues:
* #21 will fix the documentation links in the best practice list
* #23 will get rid of the string matching

Once those are done, we can make this prettier as well, but for now just
the basics is sufficient, I believe.

Before publishing, this can be run from the repository branch like `uv
run self-review --charm-name rbldnsd --repository
https://github.com/tonyandrewmeyer/rbldnsd-operator` (with any charm
repository you like). I'll also add an example output to this PR.